updated tweet_sentiment.py
I made typo on the previous commit saying "uint8" when I meant "utf-8".
For some reason I do not have the issue anymore so I deleted the
.decode('utf-8-sig) part of the code in the tweet_dict function
definitions.

I think I had problem with it initially due to other problems. But it
was solved later when decode the data correctly.